+++ worldedit-core/src/main/java/com/sk89q/worldedit/EditSession.java
@@ -532,7 +532,7 @@ public class EditSession extends PassthroughExtent implements AutoCloseable {
      * @see #getLimit()
      */
     @Deprecated
-    public int getBlockChangeLimit() {
+    public long getBlockChangeLimit() {
         return originalLimit.MAX_CHANGES;
     }
 
@@ -541,7 +541,7 @@ public class EditSession extends PassthroughExtent implements AutoCloseable {
      *
      * @param limit the limit (&gt;= 0) or -1 for no limit
      */
-    public void setBlockChangeLimit(int limit) {
+    public void setBlockChangeLimit(long limit) {
         this.limit.MAX_CHANGES = limit;
     }
